Update scoreboard operators

+++ PvpStats/Types/Display/CCAggregateStats.cs
@@ -2,12 +2,12 @@
 
 namespace PvpStats.Types.Display;
 public class CCAggregateStats {
-    public uint Matches { get; set; }
-    public uint Wins { get; set; }
-    public uint Losses { get; set; }
-    public uint OtherResult => Matches - Wins - Losses;
+    public int Matches { get; set; }
+    public int Wins { get; set; }
+    public int Losses { get; set; }
+    public int OtherResult => Matches - Wins - Losses;
     public double WinRate => (double)Wins / (Wins + Losses);
-    public int WinDiff => (int)Wins - (int)Losses;
+    public int WinDiff => Wins - Losses;
     public Job? Job { get; set; }
 
     public static CCAggregateStats operator +(CCAggregateStats a, CCAggregateStats b) {
